requirements.txt
flask
firebase-admin
mysql-connector-python
requests
python-dotenv
google-auth
